fix: 起動時の白画面を修正
- main.ts と App.svelte 間の循環インポートを解消（waitForSwCheck を独立モジュール src/lib/sw-check.ts に分離）
- テーマ適用を i18n 初期化より前に移動し、ローディング画面に正しい背景色を反映
- HomeView にデータ取得中のローディングインジケータを追加（isFirstPriorityFetched が false の間）
